refactor: rename utils about coins directory
