+++ src/library/storage/vendingMachineCoins.js
@@ -1,8 +1,8 @@
-import addCoins from '../../machine/utils/addCoins.js';
-import calculateSumOfCoins from '../../machine/utils/calculateSumOfCoins.js';
-import generateRandomCoins from '../../machine/utils/generateRandomCoins.js';
-import generateZeroCoins from '../../machine/utils/generateZeroCoins.js';
-import subtractCoins from '../../machine/utils/subtractCoins.js';
+import addCoins from '../../machine/utils/coins/addCoins.js';
+import calculateSumOfCoins from '../../machine/utils/coins/calculateSumOfCoins.js';
+import generateRandomCoins from '../../machine/utils/coins/generateRandomCoins.js';
+import generateZeroCoins from '../../machine/utils/coins/generateZeroCoins.js';
+import subtractCoins from '../../machine/utils/coins/subtractCoins.js';
 import { getItem, setItem } from './utils/index.js';
 
 const KEY_VENDING_MACHINE_COINS = 'vending-machine-coins';
