+++ src/machine/utils/coins/addCoins.js
@@ -1,4 +1,4 @@
-import { COINS } from '../const.js';
+import { COINS } from '../../const.js';
 
 function addCoins(oldCoins, newCoins) {
   return COINS.reduce(
