+++ src/machine/utils/coins/calculateReturnedCoins.js
@@ -1,4 +1,4 @@
-import { COINS } from '../const.js';
+import { COINS } from '../../const.js';
 import generateZeroCoins from './generateZeroCoins.js';
 
 function calculateReturnedCoins(machineCoins, consumerCoin) {
