+++ src/machine/utils/coins/calculateSumOfCoins.js
