+++ src/machine/utils/coins/generateRandomCoins.js
